Movidle Movie Guessing Game
This is a movie guessing game. Each of the movies on IMDB's top 250 has features. If there is a similarity between the movie you guessed and the movie that the game holds, it is colored green, otherwise it is colored red. You have 5 guesses and it is aimed to find the desired movie in this framework.